Intro Battle, Explanation for locked 252, Update on Attack
The intro battle has a text to tell you how to fight the enemy (and turns the phone on automatically to make sense). I explained the player cannot enter 252 unless they go to the game design lab first. Lastly I switched how attack works. The player would not see the decrease in health until the next turn. I move it so that it decreases then prints enemy health.